dom/broadcastchannel/tests/test_broadcastchannel_worker_alive.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM BroadcastChannel in workers
-->
<head>
  <title>Test for BroadcastChannel in workers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" language="javascript">

function runTests() {
  var id = 0;
  (new BroadcastChannel('foobar')).onmessage = function(event) {
    info("MSG: " + event.data);

    if (event.data == "READY") {
      ok(true, "Worker is ready!");
    } else {
      is(id, event.data, "The message is correct: " + id);
    }

    for (var i = 0; i < 3; ++i) {
      SpecialPowers.forceCC();
      SpecialPowers.forceGC();
    }

    if (id == 5) {
      SimpleTest.finish();
      return;
    }

    event.target.postMessage(++id);
  };

  new Worker("broadcastchannel_worker_alive.js");
}

SimpleTest.waitForExplicitFinish();
runTests();

</script>
</pre>
</body>
</html>
